+++ pidController.c
@@ -46,9 +46,9 @@ void initController(Controller* controllerPointer, uint32_t K_P, uint32_t K_I, u
 // Duty cycle limits are set for altitude and yaw so as
 // to not overload the helicopter rig and emulator.
 //******************************************************
-int8_t getControlSignal(Controller *piController, int16_t reference, int16_t measurement, bool isYaw)
+int16_t getControlSignal(Controller *piController, int16_t reference, int16_t measurement, bool isYaw)
 {
-    int8_t  dutyCycle;
+    int16_t  dutyCycle;
     int16_t controlSignal;
     int16_t errorSignal;
     int32_t derivativeError;
